added some route protection

+++ routes/comments.routes.js
@@ -1,8 +1,10 @@
-const Comment = require("../models/Comment.model");
-const router = require("express").Router();
+const Comment = require('../models/Comment.model');
+const router = require('express').Router();
+const isLoggedIn = require('../middleware/isLoggedin');
+const isAdmin = require('../middleware/isAdmin');
 
 //Get all comments by user id
-router.get("/:userId", async (req, res, next) => {
+router.get('/:userId', async (req, res, next) => {
   try {
     const id = req.params.userId;
     const foundComments = await Comment.findById(id);
@@ -13,7 +15,7 @@ router.get("/:userId", async (req, res, next) => {
 });
 
 //Get all comments by place id
-router.get("/:placeId", async (req, res, next) => {
+router.get('/:placeId', async (req, res, next) => {
   try {
     const id = req.params.placeId;
     const foundComments = await Comment.findById(id);
@@ -24,7 +26,7 @@ router.get("/:placeId", async (req, res, next) => {
 });
 
 //Create comment
-router.post("/", async (req, res, next) => {
+router.post('/', isLoggedIn, async (req, res, next) => {
   try {
     const createdComment = await Comment.create(req.body);
     res.status(200).json(createdComment);
@@ -34,24 +36,30 @@ router.post("/", async (req, res, next) => {
 });
 
 //Update comment
-router.patch("/:id", async (req, res, next) => {
+router.patch('/:id', isLoggedIn, async (req, res, next) => {
   try {
     const id = req.params.id;
-    const updatedComment = await Comment.findByIdAndUpdate(id, req.body, {
-      new: true,
-    });
-    res.status(200).json(updatedComment);
+    const foundComment = await Comment.findById(id);
+    if (req.user._id.toString() === foundComment.user.toString() || isAdmin()) {
+      const updatedComment = await Comment.findByIdAndUpdate(id, req.body, {
+        new: true,
+      });
+      res.status(200).json(updatedComment);
+    }
   } catch (error) {
     res.sendStatus(404);
   }
 });
 
 //Delete comment
-router.delete("/:id", async (req, res, next) => {
+router.delete('/:id', isLoggedIn, async (req, res, next) => {
   try {
     const id = req.params.id;
-    const deletedComment = await Comment.findByIdAndDelete(id);
-    res.status(200).json(deletedComment);
+    const foundComment = await Comment.findById(id);
+    if (req.user._id.toString() === foundComment.user.toString() || isAdmin()) {
+      const deletedComment = await Comment.findByIdAndDelete(id);
+      res.status(200).json(deletedComment);
+    }
   } catch (error) {
     res.sendStatus(404);
   }
